Refactored away remaining integration tests into unit tests.
