--- Tests - Integration/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.2.1" targetFramework="net461" />
-  <package id="Castle.Windsor" version="4.1.0" targetFramework="net461" />
-  <package id="CommonServiceLocator" version="1.3" targetFramework="net461" />
-  <package id="CommonServiceLocator.WindsorAdapter" version="1.0" targetFramework="net461" />
-  <package id="NUnit" version="3.9.0" targetFramework="net461" />
-</packages>
