+++ Tests - Shared/Tests - Shared.csproj
@@ -60,12 +60,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AbstractUnitTestBase.cs" />
-    <Compile Include="TestData\PokerHandComparisonTestCaseData.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestData\CardTestCaseData.cs" />
-    <Compile Include="TestObjects\ComparableObject.cs" />
     <Compile Include="TestObjects\FakeRandomNumberGenerator.cs" />
-    <Compile Include="TestObjects\TestEnum.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
